Progress so far
TL;DR - kinda progress but I suck. want to improve. ugly crying.

---- Rant ----
Uuh I'm sorta-kinda getting the hang when it comes to the basics of the basics
I would love to improve this, but working with html/css is way more pain and
wasted time than I ever signed to...
All in all, this goes rather smoothly purely thanks to the ultimate luck that is
for once on my side. It ugly, but it does the primitives of what it's supposed to
I say this will do. If I feel like I have some time to work on visuals and the like,
I'll definitely take it on, but, well, you never know. One thing that is a must-todo
is a refactor, given how fugly the code is at the moment
---------------